Use tabs for charting application
