+++ default/python/charting/log_parser/file_reader.py
