+++ default/python/charting/log_parser/test_tokenizer.py
@@ -1,4 +1,4 @@
-from charting.log_tokenizer import _get_token_from_line
+from charting.log_parser.log_tokenizer import _get_token_from_line
 from common.statistic_interface import StatKey
