+++ default/python/charting/tabs/widgets.py
